An alias analysis that never aliased, and a cache that forgot the name
Eleven defects across the analyses and the emitters. Each was reproduced
by running the code before the change and measured again after.

The Steensgaard analysis unified nothing. propagate() merged only the
points-to targets and never the pointers, and alias() asked only whether
the pointers had been merged, so no two distinct ids could come back
anything but NoAlias -- p = q, and p = &o alongside q = &o, both
reported NoAlias, and classCount() never fell below the number of values
added. NoAlias is the permissive answer: every consumer was free to
reorder and to drop stores across pointers that genuinely alias. The
constraint rules now follow Steensgaard (POPL'96) -- copy unifies the
two pointers, load unifies the destination with the pointee, store
unifies the pointee with the stored value -- run to a fixpoint, and
alias() also asks whether the two classes point at the same class, which
is the half `p = &o; q = &o` needs. joinWithPropagation read its two
points-to iterators after union-by-rank had swapped the ids they named,
so the surviving root could be left with no edge at all; the edges are
now read by value and re-matched to the root that survived. Pointers to
different objects still do not alias.

computeIDom appended to domChildren without clearing it. SSAPass::run
builds the dominator tree twice -- once before renaming, once for the
verifier -- so a second pass over the same function walked a tree where
every child appeared twice, and SSARename recurses: over 20 blocks the
second pass took 14.8 s where the first took 0.0 ms, and 22 blocks did
not finish in two minutes. The lists are rebuilt now, as domFrontier
already was.

The ABI artifact marker never checked that the instruction it was about
to call bookkeeping wrote the stack pointer. Its own comment called it a
heuristic: "if any use is an Immediate and this is the entry block, it's
likely SUB RSP, N". It also describes `eax = ecx - 32`, and
DeadPropagation refuses to propagate liveness through an artifact, so
the computation went out with the prologue -- measured with the return
reading its result. The AND marker had the same gap, and negated
INT64_MIN to test its magnitude: at -O0 that marked `AND rsp, INT64_MIN`
as a 16-byte alignment, at -O2 it marked nothing, because the negation
is undefined and gcc folds the branch. Callee-save restores were matched
by stack offset alone, so a reload of a different register from a reused
slot was paired with the save and both deleted.

A warm decompile disagreed with a cold one, which is what CI's CACHE-05
caught: "[RetDec] unknown detected" where the cold run said "[RetDec]
std::find_if detected". The function-analysis cache writes kindName()
-- "std::find_if" -- and read it back through its own list of five bare
names, "Find", "Transform", that kindName() never produces. No kind ever
survived the round trip, and twelve of the seventeen had no case at all.
The container path had the identical drift, with eight of fourteen
missing. Both now read one table in both directions, and a cache entry
that cannot name its kind is a miss rather than a wrong answer.

GlobalVarInfo::width is a bit count computed as memWidth * 8, held in a
uint8_t: a 32-byte ymm access recorded 0, and 0 is the value the
conflict test treats as "not seen yet", so the widest accesses were both
mis-sized and exempt from the ambiguity they should have raised.

In the emitters: slashToDot stripped a leading L from every name, so
List came out ist and Locale ocale -- JVMS 4.3.2 pairs the L with the
';', and only a name carrying both is a descriptor. emitMethodCall built
its argument list from args.begin() + 1 without checking that args had
anything in it, which on an empty expression stack is a length_error
nobody catches. Every C# event was emitted as `event void` because
EventGroup::type had no assignment. C# \x takes one to four hex digits,
so "\x01" followed by an 'F' was read back as one character; it is \u
now. F# arrays recursed on the printed string and passed a
default-constructed BcType, rendering every array as unit[].

check_std_includes.sh scanned raw text, so a comment naming an algorithm
counted as calling it. Comments and string literals are blanked first;
six files left the baseline, one of them a std::remove inside a
commented-out function.

ccache reported 3 hits in 3041 calls with 92 cleanups at 2.0/2.0 GB --
the cache filled during the build and evicted each object before
anything reused it. 8G.

29 regression tests. Each fails, hangs, or aborts without the change it
covers, except four written to show that a fix did not cost the case it
was meant to keep.

+++ include/retdec/ipa/ipa.h
@@ -210,7 +210,11 @@ private:
 struct GlobalVarInfo {
     std::string name;
     uint64_t    address    = 0;
-    uint8_t     width      = 0;
+    /// Access width in BITS. Wide enough for a 512-bit AVX access: as a
+    /// uint8_t, `memWidth * 8` truncated to 0 for every access of 32 bytes or
+    /// more, so a ymm/zmm global recorded width 0 and, being 0, was then
+    /// skipped by the conflict test that would have flagged it as ambiguous.
+    uint16_t    width      = 0;
     bool        isFp       = false;
     bool        isPointer  = false;
     bool        isAmbiguous= false;  ///< conflicting types from different writers
